1、新增数据源，同时存在两个数据源 2、代码结构重构 3、打包方式修改

+++ src/main/java/com/sixku/mysql2kafka/task/canal/RowData2Bean.java
@@ -90,7 +90,7 @@ public class RowData2Bean {
                     field.set(t, Double.valueOf(rowValue));
                     break;
                 case "BigDecimal":
-                    field.set(t, BigDecimal.valueOf(Long.valueOf(rowValue)));
+                    field.set(t, BigDecimal.valueOf(Double.valueOf(rowValue)));
                     break;
                 case "Date":
                     try {
